a fix regarding interaction distance
For some interactable objects, using transform.position as the center of an object to measure distance from doesn't work, even though the pivot was on the right spot. I never fully understood why this is the case, but a working replacement is using the center of one of the meshes in the object as a point of reference.